Disable fait amount when ARK=0
Earlier, on loading the profile, if ark balance was zero, it shew 'A 0 $'(didnt show amount in $). This was because on loading profile, fiatCurrency didn't initialize.
Since it is more appropriate not to show the fiat balance when user has 0ark and this also solves the above issue, I have disabled fiat amount when the ark is 0 by using *ngIf directive.